Make the age-tracker a child of #ui-bottom (#144)
Overrides _insertElement to insert into the ui-bottom container, removes
inline positioning from the element.  Enable pointer-events for the
element and position the element 15px above the hotbar regardless of
whether its folded or not.  Removes duplicate id in the age-tracker
template.

This fixes an issue where the container for the age tracker covered a
large portion of the canvas, preventing mouse interaction in those
areas.